cdk/requirements.txt
astroid==2.4.2
attrs==19.3.0
aws-cdk.assets==1.45.0
aws-cdk.aws-autoscaling==1.45.0
aws-cdk.aws-autoscaling-common==1.45.0
aws-cdk.aws-certificatemanager==1.45.0
aws-cdk.aws-cloudformation==1.45.0
aws-cdk.aws-cloudwatch==1.45.0
aws-cdk.aws-ec2==1.45.0
aws-cdk.aws-eks==1.45.0
aws-cdk.aws-elasticloadbalancing==1.45.0
aws-cdk.aws-elasticloadbalancingv2==1.45.0
aws-cdk.aws-events==1.45.0
aws-cdk.aws-iam==1.45.0
aws-cdk.aws-kms==1.45.0
aws-cdk.aws-lambda==1.45.0
aws-cdk.aws-logs==1.45.0
aws-cdk.aws-route53==1.45.0
aws-cdk.aws-s3==1.45.0
aws-cdk.aws-s3-assets==1.45.0
aws-cdk.aws-sns==1.45.0
aws-cdk.aws-sqs==1.45.0
aws-cdk.aws-ssm==1.45.0
aws-cdk.cdk-assets-schema==1.45.0
aws-cdk.cloud-assembly-schema==1.45.0
aws-cdk.core==1.45.0
aws-cdk.custom-resources==1.45.0
aws-cdk.cx-api==1.45.0
aws-cdk.region-info==1.45.0
aws-ec2==1.0.1.dev20180715112653
boto3==1.7.4
botocore==1.10.84
cattrs==1.0.0
constructs==3.0.3
docutils==0.16
importlib-metadata==1.6.1
importlib-resources==2.0.0
isort==4.3.21
jmespath==0.10.0
jsii==1.6.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
publication==0.0.3
pylint==2.5.3
python-dateutil==2.8.1
s3transfer==0.1.13
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
wrapt==1.12.1
zipp==3.1.0
